feat: goalPercent float타입으로 변경 / plan 저장 구현

+++ src/main/java/com/example/bo/plan/dto/PlanDto.java
@@ -22,7 +22,7 @@ public class PlanDto {
     private String startDate;
     private String endDate;
     private String planName;
-    private int goalPercent;
+    private float goalPercent;
     private int oldGoalCount;
     private int newGoalCount;
     private List<Bible> goalStatus;
